+++ src/bandersnatch/mirror.py
@@ -54,7 +54,9 @@ class Mirror:
         self.altered_packages: Dict[str, Set[str]] = {}
 
     async def synchronize(
-        self, specific_packages: Optional[List[str]] = None
+        self,
+        specific_packages: Optional[List[str]] = None,
+        sync_simple_index: bool = True,
     ) -> Dict[str, Set[str]]:
         logger.info(f"Syncing with {self.master.url}.")
         self.now = datetime.datetime.utcnow()
@@ -84,7 +86,7 @@ class Mirror:
             )
 
         await self.sync_packages()
-        self.finalize_sync()
+        self.finalize_sync(sync_index_page=sync_simple_index)
         return self.altered_packages
 
     def _filter_packages(self) -> None:
@@ -164,7 +166,7 @@ class Mirror:
             # TODO Remove this check by following packages_to_sync's typing
             self.on_error(e)
 
-    def finalize_sync(self) -> None:
+    def finalize_sync(self, sync_index_page: bool = True) -> None:
         raise NotImplementedError()
 
     def on_error(self, exception: BaseException, **kwargs: Dict) -> None:
@@ -341,8 +343,9 @@ class BandersnatchMirror(Mirror):
         # Cleanup old legacy non PEP 503 Directories created for the Simple API
         await self.cleanup_non_pep_503_paths(package)
 
-    def finalize_sync(self) -> None:
-        self.sync_index_page()
+    def finalize_sync(self, sync_index_page: bool = True) -> None:
+        if sync_index_page:
+            self.sync_index_page()
         if self.need_wrapup:
             self.wrapup_successful_sync()
         return None
@@ -979,7 +982,9 @@ class BandersnatchMirror(Mirror):
 
 
 async def mirror(
-    config: configparser.ConfigParser, specific_packages: Optional[List[str]] = None
+    config: configparser.ConfigParser,
+    specific_packages: Optional[List[str]] = None,
+    sync_simple_index: bool = True,
 ) -> int:
     config_values = validate_config_values(config)
 
@@ -1046,7 +1051,9 @@ async def mirror(
             download_mirror=config_values.download_mirror,
             download_mirror_no_fallback=config_values.download_mirror_no_fallback,
         )
-        changed_packages = await mirror.synchronize(specific_packages)
+        changed_packages = await mirror.synchronize(
+            specific_packages, sync_simple_index=sync_simple_index
+        )
 
     logger.info(f"{len(changed_packages)} packages had changes")
     for package_name, changes in changed_packages.items():